Create a separate test class for 410 Gone tests
